major refactor of ether, liaison, manager, session, added discovery class
