+++ src/ether/_internal/_pubsub.py
@@ -1,14 +1,43 @@
 import zmq
 import uuid
 from typing import Optional
+import os
+import signal
 
 from ..utils import get_ether_logger
-from ..config import EtherNetworkConfig
-from ._session import EtherSession
 
 # pubsub message frame indices
 PUSUB_MSG_TOPIC_INDEX = 0
 PUSUB_MSG_DATA_INDEX = 1
+
+def _run_pubsub(
+        frontend_port: int,
+        backend_port: int,
+):
+    """Standalone function to run PubSub proxy
+    
+    Args:
+        frontend_port: Port for the XPUB socket
+        backend_port: Port for the XSUB socket
+    """
+
+    
+    proxy = _EtherPubSubProxy(
+        frontend_port=frontend_port,
+        backend_port=backend_port
+    )
+    
+    def handle_stop(signum, frame):
+        """Handle stop signal by cleaning up proxy"""
+        proxy._logger.debug("Received stop signal, shutting down proxy...")
+        proxy.cleanup()
+        os._exit(0)  # Exit cleanly
+    
+    # Register signal handlers
+    signal.signal(signal.SIGTERM, handle_stop)
+    signal.signal(signal.SIGINT, handle_stop)
+    
+    proxy.run()
 
 class _EtherPubSubProxy:
     """Proxy that uses XPUB/XSUB sockets for efficient message distribution.
@@ -16,20 +45,19 @@ class _EtherPubSubProxy:
     XPUB/XSUB sockets are special versions of PUB/SUB that expose subscriptions
     as messages, allowing for proper subscription forwarding.
     """
-    def __init__(self, network_config: Optional[EtherNetworkConfig] = None):
+    def __init__(
+            self, 
+            frontend_port: int, 
+            backend_port: int
+        ):
         self.id = uuid.uuid4()
         self.name = f"EtherPubSubProxy_{self.id}"
         self._logger = get_ether_logger("EtherPubSubProxy")
         self._logger.debug("Initializing PubSub proxy")
         
-        # Use provided network config or get from session
-        if network_config is None:
-            session_data = EtherSession.get_current_session()
-            if session_data and "network" in session_data:
-                network_config = EtherNetworkConfig.model_validate(session_data["network"])
-            else:
-                network_config = EtherNetworkConfig()
-        self.network = network_config
+
+        self.pubsub_frontend_port = frontend_port
+        self.pubsub_backend_port = backend_port
         
         self.capture_socket = None
         self.broadcast_socket = None
@@ -43,16 +71,16 @@ class _EtherPubSubProxy:
         self._zmq_context = zmq.Context()
         
         # Setup capture (XSUB) socket
-        self._logger.debug(f"Setting up XSUB socket on port {self.network.pubsub_backend_port}")
+        self._logger.debug(f"Setting up XSUB socket on port {self.pubsub_backend_port}")
         self.capture_socket = self._zmq_context.socket(zmq.XSUB)
-        self.capture_socket.bind(f"tcp://*:{self.network.pubsub_backend_port}")
+        self.capture_socket.bind(f"tcp://*:{self.pubsub_backend_port}")
         self.capture_socket.setsockopt(zmq.RCVHWM, 1000000)
         self.capture_socket.setsockopt(zmq.RCVBUF, 65536)
         
         # Setup broadcast (XPUB) socket
-        self._logger.debug(f"Setting up XPUB socket on port {self.network.pubsub_frontend_port}")
+        self._logger.debug(f"Setting up XPUB socket on port {self.pubsub_frontend_port}")
         self.broadcast_socket = self._zmq_context.socket(zmq.XPUB)
-        self.broadcast_socket.bind(f"tcp://*:{self.network.pubsub_frontend_port}")
+        self.broadcast_socket.bind(f"tcp://*:{self.pubsub_frontend_port}")
         self.broadcast_socket.setsockopt(zmq.SNDHWM, 1000000)
         self.broadcast_socket.setsockopt(zmq.SNDBUF, 65536)
         self.broadcast_socket.setsockopt(zmq.XPUB_VERBOSE, 1)
