+++ src/scripts/network_setup/test_ether_network_server.py
@@ -1,7 +1,7 @@
 from ether import ether
 from ether._internal._config import _EtherInstanceConfig
 from ether import ether_save, ether_get, ether_start
-from ether.config import EtherConfig, EtherNetworkConfig
+from ether.config import EtherConfig, EtherSessionConfig
 from ether.utils import get_ether_logger, get_ip_address
 import socket
 import requests
@@ -44,12 +44,12 @@ def run_server(host: str = None):
     ip = get_ip_address(use_public=True)
     logger.info(f"Server IP address: {ip}")
     
-    network_config = EtherNetworkConfig(
+    session_config = EtherSessionConfig(
         host=ip,  # Use public/local IP
     )
     
     config = EtherConfig(
-        network=network_config,
+        session=session_config,
         instances={
             "network_test": _EtherInstanceConfig(
                 class_path="tests.test_ether_network_server.NetworkTestService",
